# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
      "AccentPhrase": {
        "description": "アクセント句ごとの情報",
        "properties": {
          "accent": {
            "title": "アクセント箇所",
            "type": "integer"
          },
          "is_interrogative": {
            "default": false,
            "title": "疑問系かどうか",
            "type": "boolean"
          },
          "moras": {
            "items": {
              "$ref": "#/components/schemas/Mora"
            },
            "title": "モーラのリスト",
            "type": "array"
          },
          "pause_mora": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Mora"
              }
            ],
            "title": "後ろに無音を付けるかどうか"
          }
        },
        "required": [
          "moras",
          "accent"
        ],
        "title": "AccentPhrase",
        "type": "object"
      },
      "AudioQuery": {
        "description": "音声合成用のクエリ",
        "properties": {
          "accent_phrases": {
            "items": {
              "$ref": "#/components/schemas/AccentPhrase"
            },
            "title": "アクセント句のリスト",
            "type": "array"
          },
          "intonationScale": {
            "title": "全体の抑揚",
            "type": "number"
          },
          "kana": {
            "title": "[読み取り専用]AquesTalk 風記法によるテキスト。音声合成用のクエリとしては無視される",
            "type": "string"
          },
          "outputSamplingRate": {
            "title": "音声データの出力サンプリングレート",
            "type": "integer"
          },
          "outputStereo": {
            "title": "音声データをステレオ出力するか否か",
            "type": "boolean"
          },
          "pitchScale": {
            "title": "全体の音高",
            "type": "number"
          },
          "postPhonemeLength": {
            "title": "音声の後の無音時間",
            "type": "number"
          },
          "prePhonemeLength": {
            "title": "音声の前の無音時間",
            "type": "number"
          },
          "speedScale": {
            "title": "全体の話速",
            "type": "number"
          },
          "volumeScale": {
            "title": "全体の音量",
            "type": "number"
          }
        },
        "required": [
          "accent_phrases",
          "speedScale",
          "pitchScale",
          "intonationScale",
          "volumeScale",
          "prePhonemeLength",
          "postPhonemeLength",
          "outputSamplingRate",
          "outputStereo"
        ],
        "title": "AudioQuery",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Mora": {
        "description": "モーラ（子音＋母音）ごとの情報",
        "properties": {
          "consonant": {
            "title": "子音の音素",
            "type": "string"
          },
          "consonant_length": {
            "title": "子音の音長",
            "type": "number"
          },
          "pitch": {
            "title": "音高",
            "type": "number"
          },
          "text": {
            "title": "文字",
            "type": "string"
          },
          "vowel": {
            "title": "母音の音素",
            "type": "string"
          },
          "vowel_length": {
            "title": "母音の音長",
            "type": "number"
          }
        },
        "required": [
          "text",
          "vowel",
          "vowel_length",
          "pitch"
        ],
        "title": "Mora",
        "type": "object"
      },
      "MorphableTargetInfo": {
        "properties": {
          "is_morphable": {
            "title": "指定した話者に対してモーフィングの可否",
            "type": "boolean"
          }
        },
        "required": [
          "is_morphable"
        ],
        "title": "MorphableTargetInfo",
        "type": "object"
      },
...
      "Speaker": {
        "description": "話者情報",
        "properties": {
          "name": {
            "title": "名前",
            "type": "string"
          },
          "speaker_uuid": {
            "title": "話者のUUID",
            "type": "string"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/SpeakerStyle"
            },
            "title": "スタイルの一覧",
            "type": "array"
          },
          "supported_features": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpeakerSupportedFeatures"
              }
            ],
            "title": "話者の対応機能"
          },
          "version": {
            "default": "話者のバージョン",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "speaker_uuid",
          "styles"
        ],
        "title": "Speaker",
        "type": "object"
      },
...
      "SpeakerStyle": {
        "description": "話者のスタイル情報",
        "properties": {
          "id": {
            "title": "スタイルID",
            "type": "integer"
          },
          "name": {
            "title": "スタイル名",
            "type": "string"
          },
          "type": {
            "default": "talk",
            "enum": [
              "talk",
              "singing_teacher",
              "frame_decode",
              "sing"
            ],
            "title": "スタイルの種類。talk:音声合成クエリの作成と音声合成が可能。singing_teacher:歌唱音声合成用のクエリの作成が可能。frame_decode:歌唱音声合成が可能。sing:歌唱音声合成用のクエリの作成と歌唱音声合成が可能。",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "SpeakerStyle",
        "type": "object"
      },
      "SpeakerSupportPermittedSynthesisMorphing": {
        "description": "An enumeration.",
        "enum": [
          "ALL",
          "SELF_ONLY",
          "NOTHING"
        ],
        "title": "SpeakerSupportPermittedSynthesisMorphing",
        "type": "string"
      },
      "SpeakerSupportedFeatures": {
        "description": "話者の対応機能の情報",
        "properties": {
          "permitted_synthesis_morphing": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpeakerSupportPermittedSynthesisMorphing"
              }
            ],
            "default": "ALL",
            "title": "モーフィング機能への対応"
          }
        },
        "title": "SpeakerSupportedFeatures",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "DUMMY の音声合成エンジンです。",
    "title": "DUMMY Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/install_library/{library_uuid}": {
      "post": {
        "description": "音声ライブラリをインストールします。\n音声ライブラリのZIPファイルをリクエストボディとして送信してください。",
        "operationId": "install_library_install_library__library_uuid__post",
        "parameters": [
          {
            "description": "音声ライブラリのID",
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "description": "音声ライブラリのID",
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Library",
        "tags": [
          "音声ライブラリ管理"
        ]
      }
    },
...
    "/morphable_targets": {
      "post": {
        "description": "指定されたベーススタイルに対してエンジン内の各話者がモーフィング機能を利用可能か返します。\nモーフィングの許可/禁止は`/speakers`の`speaker.supported_features.synthesis_morphing`に記載されています。\nプロパティが存在しない場合は、モーフィングが許可されているとみなします。\n返り値のスタイルIDはstring型なので注意。",
        "operationId": "morphable_targets_morphable_targets_post",
        "parameters": [
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "title": "Base Style Ids",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "$ref": "#/components/schemas/MorphableTargetInfo"
                    },
                    "type": "object"
                  },
                  "title": "Response Morphable Targets Morphable Targets Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定したスタイルに対してエンジン内の話者がモーフィングが可能か判定する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
    "/speakers": {
      "get": {
        "description": "話者情報の一覧を取得します。",
        "operationId": "speakers_speakers_get",
        "parameters": [
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Speakers Speakers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speakers",
        "tags": [
          "その他"
        ]
      }
    },
...
    "/synthesis_morphing": {
      "post": {
        "description": "指定された2種類のスタイルで音声を合成、指定した割合でモーフィングした音声を得ます。\nモーフィングの割合は`morph_rate`で指定でき、0.0でベースのスタイル、1.0でターゲットのスタイルに近づきます。",
        "operationId": "_synthesis_morphing_synthesis_morphing_post",
        "parameters": [
          {
            "in": "query",
            "name": "base_speaker",
            "required": true,
            "schema": {
              "title": "Base Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "target_speaker",
            "required": true,
            "schema": {
              "title": "Target Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "morph_rate",
            "required": true,
            "schema": {
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Morph Rate",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "2種類のスタイルでモーフィングした音声を合成する",
        "tags": [
          "音声合成"
        ]
      }
    },
    "/uninstall_library/{library_uuid}": {
      "post": {
        "description": "音声ライブラリをアンインストールします。",
        "operationId": "uninstall_library_uninstall_library__library_uuid__post",
        "parameters": [
          {
            "description": "音声ライブラリのID",
            "in": "path",
            "name": "library_uuid",
            "required": true,
            "schema": {
              "description": "音声ライブラリのID",
              "title": "Library Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uninstall Library",
        "tags": [
          "音声ライブラリ管理"
        ]
      }
    },
...
  }
}